% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{print.palette}
\alias{print.palette}
\title{Palette Print Function}
\usage{
\method{print}{palette}(x, ...)
}
\arguments{
\item{x}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
Palette Print Function
}
