% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{whistledown_palettes}
\alias{whistledown_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 10.
}
\usage{
whistledown_palettes
}
\description{
Use \code{names(whistledown_palettes)} to view list of palette names.
Currently:  "alaquod" "durorthod"  "eutrostox" "natrudoll" "paleustalf" "podzol" "rendoll" "vitrizerand" "bangor" "pywell" "crait" "redox" "redox2" "gley"
}
\keyword{datasets}
